+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Fixed incorrect ``Apply changes first`` in the Organ Settings dialog on OSX https://github.com/oleg68/GrandOrgue/issues/87
 - Fixed size of the Organ Selection dialog https://github.com/oleg68/GrandOrgue/issues/85
 - Added asking for restart after changing language (does not work on Windows yet)
 - Added displaying a reason of entering to the settings on GO startup https://github.com/oleg68/GrandOrgue/issues/66